crossplane: fix broken build
